Removing links, properly initializing dialogs
Removing the <a xlink:href>s from the svg file. They would probably be
fine after the dialog got working but there is no need for them to be
in the file if it they aren’t going to be used.

Changing the JS on how the dialogs work because, after I read the
documentation, I was doing it wrong. Now they initialized, told not to
open and have a separate function to open the dialogs.