Advanced search options on directory page (#2)
* Attached terms attempt #1

Trying to get the terms list for each entry so they can be used as
classes to show/hide them on the directory page so we can filter by
department

* Attached terms attempt #2

Made more visible changes to make success/failure more visible.
Specifically added “advanced options” as text at the top of the page.
Also attempting to sanitize the department string so it can be used as
a class around the <li> container

* Revert "Attached terms attempt #2"

This reverts commit 2c6fe3d9f5bdf4791f64e40f9ca769d7b72f444f.

* Using the taxonomy code from the ACF documentation to get the directory
entry department to show up as a class for the <li>

* Attached terms take #3

More slight variations on the getting the attached departments to
display on the directory page

* Revert "Attached terms take #3"

This reverts commit 225da9f43006046e36d5abe80c58f4c89e66df47.

* Trying to fix broken page

Is the PHP broken or is this a caching thing?

* I can get the term links using: get_the_term_list

Trying to to find a way to easily separate the term from the link via
some code I found on the WordPress developers website

* Progress!

Previous attempt worked to get the list of terms for each directory
entry. Now trying to use those as class names on the <li> container and
trying to get a list of ALL the terms in the department taxonomy with
checkboxes at the top of the page

* Initial JS

Styling the label/checkbox columns and taking an initial stab at some
JavaScript to show/hide the departments

* JS did not work… Pretty sure the culprit is bad syntax between the
jQuery selector and the variable. Correcting that to see if it makes
any difference

* Correcting bad CSS selectors

* Tweaking the show/hide department JS.
Adding (well, attempting) to add a “reset department” button that shows
everything again

* Trying to figure out why the browser doesn’t like jQuery("#az-slider
li."+departmentValue).show(); by removing the #az-slider

* .val() might not be the right way to get the checkbox value, trying
.attr(“value”)

* Changing checkboxes to a select list and refactoring JS to try and get
the show/hide functionality to work

* The select list worked so much better than the checkboxes. Refining
that and cleaning up the styling before showing it to other people.

* Attempting to hide the department filter unless the page is /directory

* Removing redundant JS (function fixAZListingScroll) that kept the
scrolled to letter from moving behind the fixed nav bar - this is
converted by JS in functions.js

Added a “resetFilter” that allows users to easily undo the department
filter

* Correcting bad selector in the stylesheet and adding the greenButton
class to the reset filter button

* Fixing reset filter button

Correcting styling on the reset filter button so that it is a little
wider and on the same line as the select list

* Correcting text on the “show everything” section of the department
filter

* Hide reset filter button

Hiding the reset filter button unless the filter has been activated

* Hiding the reset filter button on load via an inline style. This should
have been included in revision cbb93f3

* Now that the reset filter button appears and disappears the margins
were a little off… fixing those

* Enqueuing feedback2.js

Enqueuing feedback2.js so the JotForm modal window works with fewer
steps

* Correcting previous commit

Enqueued the previous script wrong in the previous commit (eee1f55)

* Removing the feedback2.js

This didn’t work and I don’t feel like messing with it anymore so the
commit is being reverted

* Correcting bad reversion

Did not revert that correctly and it broke everything. Manually
deleting the wrong lines

* Changing the name links to use <h4> elements with a class instead of
<strong> so it will be a little more accessible

* Cleaning up after pull request was created

Removing empty <style> tags and changing the no-content message at the
end of the if statement

* Debugging some JS weirdness in the select list show all departments
listing

* Fixing the logic on the show all select department JS