Adding script from Springshare/Libcal
Adding script from Springshare/Libcal that was sent from the Library on
the page.

Using the JavaScript option from the code that was sent to me.
Attempting to enqueue the JS file and using the HTML container in the
section. Removed the PHP function that used to do this
(get_todays_library_hours) because it will not be used