Customizing single event in list view
Customizing the single event entry in list view. The goal is to use the collapse-o-matic shortcode to condense the event excerpt on list-view pages so they take up less space when looking at the view on smaller screens. The primary target of this is the link from the mobile app