Collapse events list view cleanup (#31)
Squash/Merge the collapse-events-list-view-cleanup branch with master. No idea how things got this tangled but this is the branch created by cherry picking commits from the collapsible-events-list-view branch that somehow got tangled with the embedded-ellucian-form branch

* Customizing single event in list view

Customizing the single event entry in list view. The goal is to use the collapse-o-matic shortcode to condense the event excerpt on list-view pages so they take up less space when looking at the view on smaller screens. The primary target of this is the link from the mobile app

* testing change

Making an obvious change to see if the correct template is being loaded

* Styling collapsed events

Problem with previous commit was caching. Removing hardcoded cachebusting code, clarifying text and removing the default styles

* Correcting typo and removing character

Spelling "Click" correctly and removing the down arrow because it didnt add anything. Adding a bottom border to the events to provide some definition

* Trying to force the divider bottom

Trying to force the divider on the tribe events list by adding !important

* Correcting typo

Previous commits were using the wrong selector: the correct is .type-tribe_events (underscore not hyphen)

* Tweaking link text

Trying to change the link text so that there is some differentation between the expand for info and signup

* Fighting with quotes

PHP does different things with single and double quotes and I think I used the wrong ones in the previous commit

* Not using FontAwesome on the events list

Using FontAwesome icons is difficult so we aren't going to do that right now. Maybe something will come to me once I'm home

* Editing event link text

Editing event link text and adding an icon

* Correcting typo

Correcting typo in the code

* Adding some accessibility code to the Font Awesome icons that are now being used on the single list view. Trying to replace the plugins (the events calendar) tribe_events_get_the_excerpt() function with get_first_paragraph() to eliminate some of the code that appears on the list