The logic behind the previous commit was flawed: it worked fine once the page had been saved in WordPress, but until the page had been saved it there was no value for the field which is equivalent to the saved value for hide so it was not showing up. Trying to correct that by creating a variable ($showRequestInfo) and setting it to 1, then overriding the value of the variable with the value from the field.
Since this is no longer one line of code, the PHP that controls the request info button has been moved to a new file (template-parts/request-info-button.php) and all the templates load that file. This will allow us to make changes in one place instead of several places which will make maintenance easier.