Big commit - starting to the polishing process so this piece of staging coal can become a production diamond.
Removed the original events code and styles from home-page-content.php and styles.css. The code in home-page-content.php had been commented out in a previous commit but now has been removed entirely. The rules in style.css were there as a reference and were not in use (the new code had "2" as a suffix to each class). NOTE - the classes and rules should be changed before this goes live.

style.css also had rules added to it so that the read more link in the event posting showed an icon after the text and a rule to add a small amount of padding to .listText p elements.

There were a lot of changes to functions.php:
Removed the original render_homepage_event function in favor of the the display_homepage_event function I created for this project.
Changed the container from <div class=event2> to <a class=event2> so that the entire event listing is a clickable link to the event.
Changed the get_first_paragraph function to strip out links and bolded elements from the output.