Cleaning up the news/events code before it gets to master.
Removed the "2" suffix from the events classes because the original has been removed from the site and that is no longer needed.
Adding the production events snippet to the homepage to see what will happen. The issue is that the id for the staging shortcode and the production shortcode will be different and we're trying to minimize downtime going forward. The goal is that WordPress ignores the shortcode it doesn't recognize so nothing breaks.