Adding index
After the campus map meeting on Thursday we decided on putting in a section on the page to provide some context for what the buildings are. Putting that divider in and breaking the styles for this page out into their own stylesheet so we aren't bogging down the rest of the site with all these custom styles.