Gallery block is displayed
Tip: when you do things right, things work properly.

Previous issue was that I was loading the wrong page id so the images weren't being loaded. Now that that is figured out, trying to modify the carousel/slider JS to affect the rendered block instead of having a row of images