Trying to get two carousels at once
Trying to get two carousels from the Berchmans and library pages to display in the index at the same time. This is being attempted by modifying the loop to pull content from two page ids and pulling the pertinent bits of content. Also tweaking the title field to remove "Private " from the beginning