Big open: initial modal attempt
Since carousel/slideshows seem to be working on the page, lets try to get them working inside a modal. Splicing in the micromodal setup that is being used for the Ellucian form into this. Iterating over the repeater field twice: once to get the modal buttons that will trigger the opening (this is just a demo, once this works it works this functionality will move to the buildings on the map) and once to create the modals with the carousels inside of it. This commit has MANY moving pieces and I'm pretty sure it will not work on the first try. Having said that, goals are this:
1.Have a button for the two items with field data (Berchmans and library) show up on the page