An issue: duplicated ids
The modals are not opening (modaling?) because of duplicated ids. I created a short name (berchmans instead of Berchmans Hall) to use as the id for these modals because it wouldn't have spaces or be weird. However, I had already used those ids in the SVG to identify these buildings. Removing the id attribute from the buildings in the SVG to continue testing.