Undoing some of the modal
Undoing part of the modal because it will not be used. Realized halfway through the last commit that we decided not to put a modal over the map because it hide the location on the map. This functionality will have to be recreated in the index/right side and then tweaked to work on a small screen