Changing SlickJS initialization
The SlickJS carousel is the wrong size because it starts off as hidden and the library can't deal with that. The solution seems to be to initialize the carousel when the accordion is clicked and destroy it when the accordion collapses. The way I'm doing this is pretty heavy handed and there will probably be a performance hit but I'm only trying to test a theory with this attempt