Partial success
Previous commit didn't fail but it wasn't successful either. The problem stemmed from the fact that my initial selector wasn't specific enough so I grabbed the ID of the svg (Layer_1) instead of the id of the building. Changing all the polygons to paths and the select to "svg path" to see what happens.