Delay problem is selector based
The "things will only close" issue isn't because of a the JS its because of the selector. When a user clicks the option, the JS is getting the id of the building (berchmans, library) and doing the correct thing but then it is also getting the id of the all-encompassing <g> element (campusMap) and closing things again. Trying to remove the all-encompassing <g> element to see if that works with the zoom.