Adding container and changing styles
Adding a container (#slideshowContainer) around the slideshows to elminate the excess width issue in Chrome. This might require me to refactor the JS but I'm not sure yet.
Cleaned up the class names so they are original and not holdovers from the failed accordion experiment.
Added a console.log message to ensure we're getting the latest version of the page