Functionizing the show/hide JS
Changing the JS that controls the visibility of the building slideshow from random JS into functions. This will make maintenance and changes easier.

In case it wasn't in a previous commit: https://github.com/avcs06/SVGPanZoom is the definitive link to the JS library I'm using to control the <svg>