JS Fix 2 - variable scope is the devil
Pretty sure that the functionalizing failed because variables weren't being passed to the functions and $this was getting lost (two separate problems). Fixing that my removing the showSlideshow function and putting that code back into the click function. I'm so confident in this approach that I've removed the console messages