Updates encountered while removing the disparate request info forms.
Removing template (WordPress name Nursing Interior Landing Page) because it was not being used.
Styling tweaks to make the request info button look similar to other buttons in the left rail. And a display override to make it appearin the body of a page no matter the width of the browser window