Attempting to center the map on the clicked building.
Using some code that I found on StackOverflow to try and center the map on the clicked building. Expectation for this test: intense zooming on the building and the rest of the map elements will disappear. If this works, I'll have to find a way to make it work with the pan/zoom library that I'm using.