Updating map.
Adding line breaks to the embedded svg to make it easier to read. Also removed the <title> field on the main <svg> in favor of added <title> tags to each path/polygon element so they show up when a user hovers over them.
Changed the pointer style so that the cursor only changes to a pointer when the user hovers over an element with a slideshow