Creating a new template - "No special fields (blocks editor)"
We're currently in a weird state of flux between the blocks/Gutenberg editor and the original WordPress editor. Any page that requires special fields (like the program pages) cannot use the blocks editor because the field content will be lost. However, there are lots of pages that are text-based or includes simple constructs (like tables, images, accordions) that could be moved to the blocks editor. The goal of the new template is twofold: provide an easily identifiable way to show a page doesn't require special fields and give us a way to identify content that should be using the blocks editor.