Updating the map to hold descriptions.
Created an ACF field to hold descriptions so they can be added/edited from the admin screen. Added these fields to the array so they will be printed in the slideshow. Added the two sports fields to the svg.
Added the original version of the svg to the repository so we always have a backup