Moving styles out of the snippet
While the flip-card version of /academics was being prototyped, the styles lived in the example snippet. Now that the design has been fleshed out a bit, those styles are being moved into style.css. Hopefully not all of these will be unique so they can be consolidated with other style rules.