Move Advertising Landing Page content to codebase
Moving the advertising landing page (/lp/) content type from living in the admin pages to the codebase. This should make it harder to accidentally make a change which makes this content type harder to change.

This change also allows us to specify a block template to use for the Gutenberg/blocks editor. That should allow us to set default blocks for the content type which means that we will have an de facto template so new content will have the same blocks.