Updating campus map with correct street name
When the campus map went live, it incorrectly had "Dresser Ave" on the map instead of "Grant Street." Correcting this with a new image. Tried to add the .psd file that generates the image but the file was too large for GitHub so that will live on a shared  network drive.