Cleaning up code as project winds down
Cleaning up the repository as this project starts to wind down.

Removing js/feedback2.js because we don't use the JotForm popup functionality anymore so this is taking up space, the shortcode repsonsible for this has been deactivated on staging also.
Removing single-student_story.php because it was the same as single.php and the WordPress template hierarchy *should* load that file instead.
Updated single.php to show that is also being used for student stories.
Updating a style (style.css) to make the header work a smidge better in Chrome.