Correcting a ridiculous value
When the site was moved to WordPress, various line-height values were set at 1.55555556 - that is a ridiculous number of decimal places. Chrome was using all of them to set the value, Firefox was rounding up after four decimal places. Correcting that to use a more reasonable number of decimal places to see if that helps the alignment issue I'm troubleshooting.