Commencement subsite is 100% blocks editor
The commencement subsite is 100% blocks editor now! Huzzah! <there was much rejoicing>
Since the classic editor isn't being used anymore, we don't the template or the styles to make YouTube videos responsive - so they are being deleted.
Adding a function so that embedded items are responsive. This will probably cause a conflict when this code is combined with the "second-attempt-at-redoing-the-menu-bar-" branch but that is a problem for the future