Correcting landing page menu
Some styling was lost in the interior landing page (all iterations) where the menu below the image was centered instead of evenly spaced. Trying to fix this lead to the discovery of some styles that could be done better with flexbox instead of (table and table-cell). Making that change to simplify and modernize styling while keeping the appearance the same as the live site