Menu styling and functionality tweaks
This commit is all about the menus. The menus on interior pages (desktop) are no longer centered. Adding a rule to fix that and removing an unnecessary space from that entire set of rules.

Trying a different tactic on the left nav -> small screen menu conversion. All of the iterations yesterday were effective but they were affecting the hidden <select> not the visible <div>. Trying to iterate over the attributes in the <select> list to copy them to the <button> with the same value. Testing in console leads me to believe that this is will be successful.