Update directory.css
Found a truly bizarre bug. The AZ Listing plugin CSS file seems to be only loading in Firefox, when logged into the site via an admin account. In every other instance, the site is trying to load an .scss file which cannot be parsed. This is causing the directory index to be unstyled on the directory pages. Hotfix: manually load the correct CSS via @import on stylesheet we created so it is always being loaded.