Changing header image from background-image to <img>
One of the issues that we have on the website fairly frequently is that images aren't cropped correctly so they either look weird or we have to tweak and re-upload the image multiple times - this is a nuisance. I found a WordPress plugin called SmartCrop that promises to allow us to set the image focal point but it requires redoing some of the templates and functions for it to work. This commit changes the background image (SmartCrop doesn't support background-image) in favor of an <img> tag.