Update alert-bar.php to pull data from the main blog
Noticed that the coronavirus alert is not showing up on the commencement subsite. I could recreate that ACF field on the commencement site, but that would be easy. Instead, I'm trying to see if I can use the swtich_to_blog() function to get the data from the main site (elms.edu) and pipe that to the subsites.