Removing interior landing page template and related
Just finished converting all of the pages that used the interior landing page template to use the blocks version of the template. This means that the template can go away (templates\landing-page.php). It also means that everything that used the class that WordPress generated for the template (.page-template-landing-page) can be removed from JS the stylesheets.