Customizing the landing pages
Removing the default template for advertising landing pages because we won't need that anymore (functions.php).
Removing the hardcoded phone number in the footer in favor of pulling data from a newly-created ACF field (footer-advertising.php).
Giving the new display-post-types a max-width styling rule so that it doesn't take up the entire width of the page (advertising-landing-page.css). Also removing rules for Advanced Gutenberg blocks that are no longer being used.