Removing special fields from template in favor of blocks editor
Even of the page content on the website has been moved to the blocks editor that we can make moves to make the blocks editor the default editor for the page content type. There are some settings changes (Disable Gutenberg) on the site, but there are also template changes to be made.

Currently, the default template (page.php) loads all the custom fields, the blocks editor template (no-special-fields-blocks.php) doesn't because we've recreated that functionality with blocks. In order to make the default template blocks compatible we need to remove the hardcoded custom field templates because they have the potential to slow down pageload and display empty containers. I've ported the code from no-special-fields-blocks.php to program-page-content.php (and converted no-special-fields-blocks.php to use that file) so these templates are the same. If this goes well, we can reduce this code duplication.