Update page-heading.php
As part of the move to the blocks editor for the pages content type, we won't need the featured_image_cropped field that is created by the ACF Focal Point plugin. Deactivating the plugin with this code in it cause all the headers to trigger this section of the if statement which blew out the page. Removing this before the plugin gets deactivated again.