Starting to migrate the collapsible nav bar code to the appropriate files
Removing the jQuery UI accordion styles because the current code doesn't rely on the accordion widget. Adding in styles for the non-accordion based collapsible nav.

In order to collapse/expand the nav bar, a trigger is needed. This trigger doesn't link to the appropriate page (two different functions per click is a messy user experience) so the toggle link is duplicated. This is causing the duplicates to show up on menu on small screens - which is less than ideal. Changing the conversion function to exclude the toggle links which should eliminate the duplicates