Removing JS File and minor style tweaks
Removing skip-link-focus-fix.js because it was a shim for old browsers and new browsers handle this functionality without the file. Since the file doesn't do anything, it's dead weight that is slowing down the site and is being removed.

Minor style tweaks to nursing.css (removing an unused style) and commencement.css (hiding part of a nav bar).