Cleaning up functions.js
There is a lot of code in functions.js and Ryan isn't sure what all of it does. After careful examination, it looks like a couple of variables and a couple of functions can be removed. This should make pages load faster because the browser isn'texecuting the encapsulating function or storing variables that are used once - plus the entire file will be shorter/smaller so it will be transmitted and loaded faster.