Initial commit of using .scss conversion
Used Dreamweaver to setup Compass to compile .scss files into .css files. A couple of files were created automatically upon setup (ie.scss, screen.scss, etc.) but the rest of the .scss are straight copies of the .css files with the same name. Created the .scss files means that the .css were compiled - the biggest difference between the new and old .css files are the new ones have no whitespace because they are compressed.

Tweaked the config.rb file to automatically move style.css out of the /css directory because WordPress expects it to be in the root directory.

Created a .gitignore file (I was surprised there wasn't one already) explicitely to make it to the .sass-cache directory is not comitted.