Continue to implement scss
Added a new variable (elmspurple) for the purple color (#413556) that is used on sites.
Started implementing some conventions: spaces after colons in .scss files

We had two versions of dark green on the website- #115438 (the primary one that is used everywhere) and a darker version (#004731) that is closer to the official color. Replacing the darker version with the $elmsgreen variable to provide consistency across the pages.