Correcting poor layout decision
Previous commit to skins/single/default.php worked - the new div id showed up on the page.

While researching how to keep the column on the left side of the page, noticed that all the templates with a left rail had the following style rule: flex-direction:row-reverse. Removed that style and moved the left rail above main content so things will display properly without a CSS rule rewriting the content of the page. This *should* make the page load slightly faster because the browser doesn't have to rewrite the content.