Consolidating styles and JS tweaks
Combining some of the different color greys so they use the $lightgrey/$lightgreyborder variable - this will make the site more consistent and make maintenance easier.

Corrected the calendar class to public-event (hyphen not underscore) so it styles the appropriate content. Changed the mecHoverStyle mixin to change the background opacity on hover instead of adding/removing a border - this should work better with the default styling.

After I rearranged all the divs on the page templates, the templates don't work as well on small screens. The dropdown menu (which replaces the left nav bar on small screens) isn't being output in the right location. Hopefully, trying changing the jQuery function and target div will correct that.