Adding template divs to single event page
Adding the layout divs that the rest of the sites use to the single event template. This was mocked up in Inspector - hopefully I've carried everything over properly.

This is might work but it certainly has led a conflicting series of divs that is overcomplicated, over-styled, and will be a nightmare to maintain. But you have to make it work before you can make it better.