Add function_exists check to header
The header template has a function that adds a meta-tag if the page being loaded has anything to do with the Events Calendar (old version, made by Modern Tribe). This is an unguarded statement that will throw an error if the plugin is deactivated.

I'm not quite sure how to untangle the events calendar from the new events calendar but the first step is trying to make this template safer. Wrapped the function in an  if(function_exists('tribe_is_event')) to protect the site when the old calendar plugin is initially deactivated.