Comment out preloadCarousel function
On many pages of elms.edu, there is a 404 error that that I've never had time to track down. My suspicion has always been that this is related to the preloadCarousel() function so we're going to comment that out.

This increased in urgency because this error is preventing us from turning on the new automatic 404 IP blocking in the security suite.