Still refactoring styles
Found more header styles that were moved from style.scss to _header.scss.
Starting removing commented out (the original) code that will not be used anymore.
Found the bit of code that was adding .active to .nav-trigger and removing that - it was causing complications in the styling and is no longer needed.