More minor tweaks
More minor tweaks that are being committed.  The biggest change is a new SCSS variable ($thickgreen), some new nesting, and the removal of some commented out code that is no longer needed.

Normall, the amount of changes is not enough to warrant a commit but I'm about to break the header into distinct sections to make styling easier