Removing styles related to Tribe Events
Removing all of the .field-related-events styles and JS functions because they were being used in the left rail block for related events in the old calendar and are no longer needed.

Unrelated change in this commit: removing display: inline-grid because it was causing display errors in Firefox on the directory.