Replacing  hardcoded community gradients with a loop
This is a maintenance commit beause this finally compiles in something that *looks like* it should work and I need a save point.

Wes and I decided that each community college should have their own categeory and color code so they can be visually grouped together on the calendar. Rather than hardcode a bunch of gradients, Ryan decided that there had to be a way to use a loop to make the code easier to maintain by reducing duplicate code. This is being commited because it seems like it should work as is but can be improved